requirements.txt
chromadb
sentence-transformers
huggingface_hub
gradio
langchain-community
langchain-text-splitters
pypdf
python-dotenv
pandas
tqdm
